+++ packages/nimble-components/src/number-field/tests/number-field.spec.ts
@@ -1,0 +1,14 @@
+import {
+    DesignSystem,
+    NumberField as FoundationNumberField
+} from '@microsoft/fast-foundation';
+import { html } from '@microsoft/fast-element';
+import '../index';
+
+describe('NumberField', () => {
+    it('should have its tag returned by tagFor(FoundationNumberField)', () => {
+        expect(html`${DesignSystem.tagFor(FoundationNumberField)}`.html).toBe(
+            'nimble-number-field'
+        );
+    });
+});